refactor: change import token path, extract logic to use-focus-trap.ts
